refactor: move part panels
