+++ src/NewPicker/PickerInput/Popup/PopupPanel.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import PickerPanel, { type PickerPanelProps } from '../../PickerPanel';
+import PickerPanel, { type PickerPanelProps } from '../../../PickerPanel';
 import { PickerHackContext, type PickerHackContextProps } from '../../PickerPanel/context';
 import PickerContext from '../context';
 import { offsetPanelDate } from '../hooks/useRangePickerValue';
