+++ src/NewPicker/PickerPanel/DatePanel/index.tsx
@@ -9,8 +9,8 @@ import {
 } from '../../../utils/dateUtil';
 import type { PanelMode, SharedPanelProps } from '../../../interface';
 import { PanelContext, useInfo } from '../context';
-import PanelBody from '../PanelBody';
-import PanelHeader from '../PanelHeader';
+import PanelBody from '../../../PickerPanel/PanelBody';
+import PanelHeader from '../../../PickerPanel/PanelHeader';
 
 export interface DatePanelProps<DateType extends object> extends SharedPanelProps<DateType> {
   panelName?: PanelMode;
