+++ src/NewPicker/PickerPanel/MonthPanel/index.tsx
@@ -2,8 +2,8 @@ import * as React from 'react';
 import { formatValue, isSameMonth } from '../../../utils/dateUtil';
 import type { SharedPanelProps } from '../../../interface';
 import { PanelContext, useInfo } from '../context';
-import PanelBody from '../PanelBody';
-import PanelHeader from '../PanelHeader';
+import PanelBody from '../../../PickerPanel/PanelBody';
+import PanelHeader from '../../../PickerPanel/PanelHeader';
 
 export default function MonthPanel<DateType extends object = any>(
   props: SharedPanelProps<DateType>,
