+++ src/NewPicker/PickerPanel/YearPanel/index.tsx
@@ -2,8 +2,8 @@ import * as React from 'react';
 import type { SharedPanelProps } from '../../../interface';
 import { formatValue, isSameYear } from '../../../utils/dateUtil';
 import { PanelContext, useInfo } from '../context';
-import PanelBody from '../PanelBody';
-import PanelHeader from '../PanelHeader';
+import PanelBody from '../../../PickerPanel/PanelBody';
+import PanelHeader from '../../../PickerPanel/PanelHeader';
 
 export default function YearPanel<DateType extends object = any>(
   props: SharedPanelProps<DateType>,
