+++ src/PickerPanel/index.tsx
@@ -11,22 +11,22 @@ import type {
   PickerMode,
   SharedPanelProps,
   SharedTimeProps,
-} from '../../interface';
-import { isSame } from '../../utils/dateUtil';
-import { toArray } from '../../utils/miscUtil';
-import useLocale from '../../hooks/useLocale';
-import { getTimeConfig } from '../../hooks/useTimeConfig';
-import useToggleDates from '../../hooks/useToggleDates';
-import PickerContext from '../PickerInput/context';
-import useCellRender from '../PickerInput/hooks/useCellRender';
-import DatePanel from './DatePanel';
-import DateTimePanel from './DateTimePanel';
-import DecadePanel from './DecadePanel';
-import MonthPanel from './MonthPanel';
-import QuarterPanel from './QuarterPanel';
-import TimePanel from './TimePanel';
-import WeekPanel from './WeekPanel';
-import YearPanel from './YearPanel';
+} from '../interface';
+import { isSame } from '../utils/dateUtil';
+import { toArray } from '../utils/miscUtil';
+import useLocale from '../hooks/useLocale';
+import { getTimeConfig } from '../hooks/useTimeConfig';
+import useToggleDates from '../hooks/useToggleDates';
+import PickerContext from '../NewPicker/PickerInput/context';
+import useCellRender from '../NewPicker/PickerInput/hooks/useCellRender';
+import DatePanel from '../NewPicker/PickerPanel/DatePanel';
+import DateTimePanel from '../NewPicker/PickerPanel/DateTimePanel';
+import DecadePanel from '../NewPicker/PickerPanel/DecadePanel';
+import MonthPanel from '../NewPicker/PickerPanel/MonthPanel';
+import QuarterPanel from '../NewPicker/PickerPanel/QuarterPanel';
+import TimePanel from '../NewPicker/PickerPanel/TimePanel';
+import WeekPanel from '../NewPicker/PickerPanel/WeekPanel';
+import YearPanel from '../NewPicker/PickerPanel/YearPanel';
 
 const DefaultComponents: Components = {
   date: DatePanel,
